requirements.txt
asgiref==3.5.2
Django==4.0.2
django-markdownx==4.0.0b1
Markdown==3.4.1
Pillow==9.3.0
python-dotenv==0.21.0
sqlparse==0.4.3
